+++ tests/rigctl_parse.c
@@ -3791,7 +3791,7 @@ declare_proto_rig(dump_state)
                 "%"FREQFMT" %"FREQFMT" 0x%"PRXll" %d %d 0x%x 0x%x\n",
                 rs->rx_range_list[i].start,
                 rs->rx_range_list[i].end,
-                (uint64_t)rs->rx_range_list[i].modes,
+                rs->rx_range_list[i].modes,
                 rs->rx_range_list[i].low_power,
                 rs->rx_range_list[i].high_power,
                 rs->rx_range_list[i].vfo,
@@ -3806,7 +3806,7 @@ declare_proto_rig(dump_state)
                 "%"FREQFMT" %"FREQFMT" 0x%"PRXll" %d %d 0x%x 0x%x\n",
                 rs->tx_range_list[i].start,
                 rs->tx_range_list[i].end,
-                (uint64_t)rs->tx_range_list[i].modes,
+                rs->tx_range_list[i].modes,
                 rs->tx_range_list[i].low_power,
                 rs->tx_range_list[i].high_power,
                 rs->tx_range_list[i].vfo,
@@ -3819,7 +3819,7 @@ declare_proto_rig(dump_state)
     {
         fprintf(fout,
                 "0x%"PRXll" %ld\n",
-                (uint64_t)rs->tuning_steps[i].modes,
+                rs->tuning_steps[i].modes,
                 rs->tuning_steps[i].ts);
     }
 
@@ -3829,7 +3829,7 @@ declare_proto_rig(dump_state)
     {
         fprintf(fout,
                 "0x%"PRXll" %ld\n",
-                (uint64_t)rs->filters[i].modes,
+                rs->filters[i].modes,
                 rs->filters[i].width);
     }
 
@@ -3858,12 +3858,12 @@ declare_proto_rig(dump_state)
 
     fprintf(fout, "\n");
 
-    fprintf(fout, "0x%"PRXll"\n", (uint64_t)rs->has_get_func);
-    fprintf(fout, "0x%"PRXll"\n", (uint64_t)rs->has_set_func);
-    fprintf(fout, "0x%"PRXll"\n", (uint64_t)rs->has_get_level);
-    fprintf(fout, "0x%"PRXll"\n", (uint64_t)rs->has_set_level);
-    fprintf(fout, "0x%"PRXll"\n", (uint64_t)rs->has_get_parm);
-    fprintf(fout, "0x%"PRXll"\n", (uint64_t)rs->has_set_parm);
+    fprintf(fout, "0x%"PRXll"\n", rs->has_get_func);
+    fprintf(fout, "0x%"PRXll"\n", rs->has_set_func);
+    fprintf(fout, "0x%"PRXll"\n", rs->has_get_level);
+    fprintf(fout, "0x%"PRXll"\n", rs->has_set_level);
+    fprintf(fout, "0x%"PRXll"\n", rs->has_get_parm);
+    fprintf(fout, "0x%"PRXll"\n", rs->has_set_parm);
 
 #if 0
     gran_t level_gran[RIG_SETTING_MAX];   /*!< level granularity */
